pyproject.toml
[project]
name = "google-cloud-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-bigquery>=3.35.0",
    "google-cloud-secret-manager>=2.24.0",
    "google-cloud-storage>=3.2.0",
    "pandas>=2.3.1",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
]
